+++ src/Service/CodeCommit/src/Exception/InvalidSortByException.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace AsyncAws\CodeCommit\Exception;
+
+use AsyncAws\Core\Exception\Http\ClientException;
+
+/**
+ * The specified sort by value is not valid.
+ */
+final class InvalidSortByException extends ClientException
+{
+}