+++ src/Service/CodeCommit/tests/Unit/CodeCommitClientTest.php
@@ -3,15 +3,21 @@
 namespace AsyncAws\CodeCommit\Tests\Unit;
 
 use AsyncAws\CodeCommit\CodeCommitClient;
+use AsyncAws\CodeCommit\Input\CreateRepositoryInput;
+use AsyncAws\CodeCommit\Input\DeleteRepositoryInput;
 use AsyncAws\CodeCommit\Input\GetBlobInput;
 use AsyncAws\CodeCommit\Input\GetBranchInput;
 use AsyncAws\CodeCommit\Input\GetCommitInput;
 use AsyncAws\CodeCommit\Input\GetDifferencesInput;
+use AsyncAws\CodeCommit\Input\ListRepositoriesInput;
 use AsyncAws\CodeCommit\Input\PutRepositoryTriggersInput;
+use AsyncAws\CodeCommit\Result\CreateRepositoryOutput;
+use AsyncAws\CodeCommit\Result\DeleteRepositoryOutput;
 use AsyncAws\CodeCommit\Result\GetBlobOutput;
 use AsyncAws\CodeCommit\Result\GetBranchOutput;
 use AsyncAws\CodeCommit\Result\GetCommitOutput;
 use AsyncAws\CodeCommit\Result\GetDifferencesOutput;
+use AsyncAws\CodeCommit\Result\ListRepositoriesOutput;
 use AsyncAws\CodeCommit\Result\PutRepositoryTriggersOutput;
 use AsyncAws\CodeCommit\ValueObject\RepositoryTrigger;
 use AsyncAws\Core\Credentials\NullProvider;
@@ -20,6 +26,33 @@ use Symfony\Component\HttpClient\MockHttpClient;
 
 class CodeCommitClientTest extends TestCase
 {
+    public function testCreateRepository(): void
+    {
+        $client = new CodeCommitClient([], new NullProvider(), new MockHttpClient());
+
+        $input = new CreateRepositoryInput([
+            'repositoryName' => 'change me',
+
+        ]);
+        $result = $client->createRepository($input);
+
+        self::assertInstanceOf(CreateRepositoryOutput::class, $result);
+        self::assertFalse($result->info()['resolved']);
+    }
+
+    public function testDeleteRepository(): void
+    {
+        $client = new CodeCommitClient([], new NullProvider(), new MockHttpClient());
+
+        $input = new DeleteRepositoryInput([
+            'repositoryName' => 'change me',
+        ]);
+        $result = $client->deleteRepository($input);
+
+        self::assertInstanceOf(DeleteRepositoryOutput::class, $result);
+        self::assertFalse($result->info()['resolved']);
+    }
+
     public function testGetBlob(): void
     {
         $client = new CodeCommitClient([], new NullProvider(), new MockHttpClient());
@@ -74,6 +107,19 @@ class CodeCommitClientTest extends TestCase
         $result = $client->getDifferences($input);
 
         self::assertInstanceOf(GetDifferencesOutput::class, $result);
+        self::assertFalse($result->info()['resolved']);
+    }
+
+    public function testListRepositories(): void
+    {
+        $client = new CodeCommitClient([], new NullProvider(), new MockHttpClient());
+
+        $input = new ListRepositoriesInput([
+
+        ]);
+        $result = $client->listRepositories($input);
+
+        self::assertInstanceOf(ListRepositoriesOutput::class, $result);
         self::assertFalse($result->info()['resolved']);
     }